changed name of file, registered metric
